iTunes-AirPlay, System, music -> app_volume, 1.1.0
AirPlay for iTunes was added in this update. Find it in the volume
screen of the iTunes player. System was also added. Gives the ability to
click keyboard arrow keys, space bar, and change the system volume.

When changing volume for Spotify for iTunes, it controls the app volume
instead.

Also, version 1.1.0.